requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scikit-optimize>=0.9.0
PyPortfolioOpt>=1.5.4
# empyrical>=0.5.5  # Commented out due to compatibility issues with Python 3.13
scipy>=1.10.0
cvxpy>=1.3.0
pyyaml>=6.0
joblib>=1.3.0
# pyfolio>=0.9.2  # Commented out due to compatibility issues with Python 3.13
matplotlib>=3.7.0
seaborn>=0.12.0

# Additional optimization dependencies
ecos>=2.0.10
osqp>=0.6.0
scs>=3.0.0